+++ src/app/app-routing.module.ts
@@ -4,6 +4,7 @@ import { OverviewComponent } from './pages/overview/overview.component';
 import { AddPriceComponent } from './pages/add-price/add-price.component';
 import { ArchiveComponent } from './pages/archive/archive.component';
 import { DynamicsComponent } from './pages/dynamics/dynamics.component';
+import { DetailsComponent } from './pages/details/details.component';
 
 
 const routes: Routes = [
@@ -22,6 +23,10 @@ const routes: Routes = [
   {
     path: 'archive',
     component: ArchiveComponent
+  },
+  {
+    path: 'details/:id',
+    component: DetailsComponent
   },
   {
     path: '',