+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0-SNAPSHOT] - 2020-05-10
+### Added
+- Getting a price entry by ID
+- Details component for displaying a price entry by ID
+
 ## [1.1.0-SNAPSHOT] - 2020-05-10
 ### Added
 - Initial project setup